Fix problem caused by BAS extension
So far IMBAS was used as filename extension of sources written in
Imbastardizer. When BAS is used instead, and a Vim modeline is added to
set the filetype, the modeline of the target file was executed too,
causing two Vim error messages when the Imbastardizer code was reloaded.

Beside, when the output file is not specified in the command line, now
its default filename extension is TARGET.BAS instead of BAS. This
prevents problems when also the input file uses BAS as filename
extension.